Lab4 VerticalView screen created
